refactor(apps): the AppsRuntime contract sits beside its implementation
`createApps` was one 2,942-line closure and the `AppsRuntime` interface it
implements sat ~2,000 lines above it in the same file, so reading any single
verb meant scrolling between two distant halves of `runtime.ts`.

The contract and the shapes its verbs speak move to `types.ts`, and four of the
nested namespaces get their own module taking a small shared context
(`runtime-context.ts`) — the shape `interchange`/`history`/`review` already use.
`pins` alone was 315 lines inline; its orchestration is now `pins-surface.ts`,
beside the pure logic that was always in `pins.ts`.

Pure relocation: no behaviour changed, no test changed, and every type is still
re-exported from `./runtime.js` so the package surface is byte-identical.
